Colour variables, variable names, numbers menu, menu page items, menu changing, gesture handling
Variables created to hold the RGB values for readability.
All variable names changed to fit "x_y" format not "xY".
The numbers menu was created to match a T9 pad.
Another menu page was added with commands, and the placeholder page had its values entered.
The system for menu changing when a menu option was selected was implemented.
Full gesture handling has been created. Currently relies on console to show current code and a single line has been tested to print a number.